Make fetching use the groups gateway.
