+++ techcity/services/groups/gateway.py
@@ -1,4 +1,7 @@
 from __future__ import annotations
+
+from techcity.models import Group
+
 from .service import GroupsService
 
 
@@ -11,3 +14,6 @@ class GroupsGateway:
     def connect(self, service: GroupsService) -> None:
         """Connect to the groups service."""
         self._service = service
+
+    def all(self) -> list[Group]:
+        return self._service.all()
